+++ src/entries/entry.rs
