+++ src/utils/mod.rs
@@ -1,11 +1,11 @@
-mod entries;
-mod files;
+mod ops;
 mod running;
+mod sequence;
 
 use anyhow::{Result, anyhow};
-pub use entries::*;
-pub use files::*;
+pub use ops::*;
 pub use running::*;
+pub use sequence::*;
 use std::collections::HashSet;
 use std::error::Error;
 use std::io::{Write, stdin, stdout};
